+++ fiddles/bash/fiddle-0094-InitCProject/fiddles/c/template/main.c
@@ -1,0 +1,6 @@
+#include <stdio.h>
+
+int main(void) {
+    printf ("{{projectName}} compiled!");
+    return 0;
+}