+++ fiddles/bash/fiddle-0094-InitCProject/scripts/bin/c/_install.sh
@@ -4,26 +4,22 @@
 #  Specification           : N/A_____________________________________________________________________|
 #  Specification Path      : N/A_____________________________________________________________________|
 #  Author                  : brady house_____________________________________________________________|
-#  Create date             : 05/02/2016______________________________________________________________|
-#  Description             : GLOBAL VARIABLE CONFIG USED BY NATIVESCRIPT CREATE/START SCRIPTS._______|
+#  Create date             : 10/02/2016______________________________________________________________|
+#  Description             : MASTER C INSTALL FUNCTION(S)____________________________________________|
+#  Entry Point             : isGccInstalled__________________________________________________________|
+#  Input Parameters        : N/A_____________________________________________________________________|
+#  Initial Consumer        : ../fiddle-c.sh__________________________________________________________|
 # ---------------------------------------------------------------------------------------------------|
 #  Revision History::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::::|
 # ---------------------------------------------------------------------------------------------------|
-# Baseline Ver - CHANGELOG @ 201609160420
-# 11/26/2016 - See CHANGELOG @ 201610010420
+# Baseline Ver - CHANGELOG @ 201611280420
 # ---------------------------------------------------------------------------------------------------|
 
-export __NODE_VERSION__="5.1.0";
-export __PLATFORM__="ANDROID";
-export __ANDROID_EMULATOR_PROFILE__=320480QVGA;
-export __ANDROID_EMULATOR_EXE__="/Users/e13542/bin/android-sdk-macosx/tools/emulator";
 
-# -------------------------------------------------
-# TEMPLATE TYPE SWITCH
-# This variable controls the template used by
-# the fiddle.sh when initializing a new nativescript
-# fiddle. 2 options are supported:
-#   "js"  JavaScript
-#   "ng2" Angular2
-# --------------------------------------------------
-export __TEMPLATE_TYPE__="ng2";
+function isGccInstalled() {
+  installed=$(if [[ ! -e "${__C_COMPILER__}" ]]; then echo "false"; fi);
+  if [[ "${installed}" == "false" ]]
+  then
+      exit -1;
+  fi
+}